// frontend/src/api/generated/models/zoneRead.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * SpotPass Backend API
 * Restaurant reservation and management system
 * OpenAPI spec version: 1.0.0
 */

/**
 * Schema for reading zones
 */
export interface ZoneRead {
  id: string;
  /**
   * @minLength 1
   * @maxLength 64
   */
  name: string;
}